+++ neatoo/AddNewWareView.swift
@@ -9,7 +9,7 @@ struct AddNewWareView: View {
     @State private var brand: String = ""
     @State private var category: String = ""
     @State private var name: String = ""
-    @State private var number: Double = 0
+    @State private var number: Int = 0
     @State private var price: Double = 0
     @State private var priority: Int = 0
     @State private var purchaseDate: Date = {
